FORMAT : #39 divide GameService
