+++ src/game/game.service.ts
@@ -17,35 +17,6 @@ export class GameService {
   paddleWidth = 10;
   paddleHeight = 75;
 
-  invite(userId: number, targetUserId: number): number {
-    const gameRoom = this.gameRepository.createGameRoom();
-
-    gameRoom.player1 = { id: userId, isAccept: true };
-    gameRoom.player2 = { id: targetUserId, isAccept: false };
-    return gameRoom.socketRoomId;
-  }
-
-  accept(userId: number, roomId: number) {
-    const gameRoom = this.gameRepository.getGameRoom(roomId);
-
-    if (gameRoom.player1.id == userId) gameRoom.player1.isAccept = true;
-    if (gameRoom.player2.id == userId) gameRoom.player2.isAccept = true;
-    if (gameRoom.player1.isAccept && gameRoom.player2.isAccept) return gameRoom;
-    return;
-  }
-
-  cancel(userId: number, roomId: number) {
-    const gameRoom = this.gameRepository.getGameRoom(roomId);
-
-    let canceledPlayerId;
-    if (gameRoom.player1.id == userId) canceledPlayerId = gameRoom.player2.id;
-    else if (gameRoom.player2.id == userId)
-      canceledPlayerId = gameRoom.player1.id;
-    else return;
-    this.gameRepository.deleteGameRoom(roomId);
-    return canceledPlayerId;
-  }
-
   private gameLoop(
     room: GameRoom,
     onUpdate: (gameInfo: GameInfo) => any,
